Add two new config options to nvda.ini under [UIA]. "enabled" is a boolean that allows you to state whether UIA should be used (default is true). "minWindowsVersion" is a float specifying the minimum windows version (major.minor) where NVDA should be allowed to use UIA. Its default is 6.1 (Windows 7). This means that UIA by default is no longer going to be used in XP. This is due to the fact that there are just to  many freezes, crashes and other performance hits when using the UIAutomation client library available in XP with the platform update. Fixes #1437.

+++ source/UIAHandler.py
@@ -3,8 +3,10 @@ from ctypes.wintypes import *
 import comtypes.client
 from comtypes import *
 import weakref
+import sys
 import threading
 import time
+import config
 import api
 import queueHandler
 import controlTypes
@@ -233,15 +235,24 @@ isUIAAvailable=False
 
 def initialize():
 	global handler, isUIAAvailable
+	if not config.conf["UIA"]["enabled"]:
+		isUIAAvailable=False
+		return
+	winver=sys.getwindowsversion()
+	winver=winver.major+(winver.minor/10.0)
+	if winver<config.conf["UIA"]["minWindowsVersion"]:
+		isUIAAvailable=False
+		return
 	try:
 		handler=UIAHandler()
 	except COMError:
+		isUIAAvailable=False
 		handler=None
 		raise RuntimeError("UIA not available")
-	isUIAAvailable=True
 
 def terminate():
-	global handler
-	handler.terminate()
-	handler=None
-
+	global handler, isUIAAvailable
+	if handler:
+		handler.terminate()
+		handler=None
+		isUIAAvailable=False
